AP-4904: Create base controllers to DRY up code
