+++ app/controllers/providers/address_manuals_base_controller.rb
@@ -1,0 +1,24 @@
+module Providers
+  class AddressManualsBaseController < ProviderBaseController
+    def show
+      @form = Addresses::AddressForm.new(model: address)
+    end
+
+    def update
+      @form = Addresses::AddressForm.new(form_params)
+      render :show unless save_continue_or_draft(@form)
+    end
+
+  private
+
+    def address_attributes
+      %i[address_line_one address_line_two city county postcode lookup_postcode lookup_error]
+    end
+
+    def form_params
+      merge_with_model(address) do
+        params.require(:address).permit(*address_attributes).merge(location:)
+      end
+    end
+  end
+end
