+++ app/controllers/providers/correspondence_address/manuals_controller.rb
@@ -1,27 +1,12 @@
 module Providers
   module CorrespondenceAddress
-    class ManualsController < ProviderBaseController
+    class ManualsController < AddressManualsBaseController
       prefix_step_with :correspondence_address
-
-      def show
-        @form = Addresses::AddressForm.new(model: address)
-      end
-
-      def update
-        @form = Addresses::AddressForm.new(form_params)
-        render :show unless save_continue_or_draft(@form)
-      end
 
     private
 
-      def address_attributes
-        %i[address_line_one address_line_two city county postcode lookup_postcode lookup_error]
-      end
-
-      def form_params
-        merge_with_model(address) do
-          params.require(:address).permit(*address_attributes).merge(location: "correspondence")
-        end
+      def location
+        "correspondence"
       end
 
       def address
