+++ app/controllers/providers/home_address/manuals_controller.rb
@@ -1,16 +1,7 @@
 module Providers
   module HomeAddress
-    class ManualsController < ProviderBaseController
+    class ManualsController < AddressManualsBaseController
       prefix_step_with :home_address
-
-      def show
-        @form = Addresses::AddressForm.new(model: address)
-      end
-
-      def update
-        @form = Addresses::AddressForm.new(form_params)
-        render :show unless save_continue_or_draft(@form)
-      end
 
     private
 
@@ -23,16 +14,6 @@ module Providers
         )
       end
 
-      def address_attributes
-        %i[address_line_one address_line_two city county postcode lookup_postcode lookup_error]
-      end
-
-      def form_params
-        merge_with_model(address) do
-          params.require(:address).permit(*address_attributes).merge(location: "home")
-        end
-      end
-
       def address
         @address ||= current_uk_home_address || build_address
       end
@@ -41,6 +22,10 @@ module Providers
         return nil unless applicant.home_address && applicant.home_address.country_code == "GBR"
 
         applicant.home_address
+      end
+
+      def location
+        "home"
       end
     end
   end
